requirements.txt
amqp==5.0.9
asgiref==3.4.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0
django-filter==21.1
django-redis-cache==3.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
environs==9.3.5
gunicorn==20.1.0
kombu==5.2.2
marshmallow==3.14.1
Pillow==8.4.0
prompt-toolkit==3.0.24
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
